feat: create custom hook `useControls`
